Move award flow views into a separate module
'outcome' seems a better name than 'award or cancel or unsuccessful'...
